Flattened file structure, and improved the config file setup.
Now there is a module that main sets up with the config informaton, and
anyone can access it.
It could be improved some with some minor changes, but the code that
uses it is better. I also removed that damn absolute path for the
botfood folder, though it should now be a config option I guess.