Change the bot entry in the config file to be a list
Having to create a dictionary of bots in the config file was not user
friendly. Creating a list makes more sense. I added code in config.py to
turn the list into a dictionary so it now gets the best of both worlds.

I also added more documentation about the fields in the config.yaml.